# core/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: Wachbuch API
  description: |
    Versionierte JSON-API für selbst gehostetes Wachbuch.
    Widerrufbare App-Tokens (`wb_…`) und stationsbezogene Rechte.
    Wachalltag umfasst Schichtübergaben, Mängel, Fahrzeug-/Gerätestatus,
    Checklisten, Schlüssel-/Pool-Geräte, Quittierungen und leichte Auswertung.
    Zusätzliche, je Station aktivierbare Module: Pinnwand (`/pinnwand/`) sowie
    der Ende-zu-Ende-verschlüsselte Nachrichtenbereich aus Wachenchat,
    privaten Zweiergesprächen, Gruppenräumen und Secure Mail
    (`/chat/*`, `/post/*`). Der Server speichert und liefert dabei nur
    Ciphertext und Schlüsselumschläge aus; er kann Nachrichteninhalte nicht lesen.
    Patienten-, Einsatz-, Alarmierungs- und Vorgangsdaten sind ausdrücklich
    nicht Teil dieser API.
  version: "1.4.0"
servers:
  - url: /api/v1
    description: Relative to the Wachbuch origin
security:
  - TokenAuth: []
paths:
...
  /me/:
    get:
      summary: Current user, membership, station modules and Wachalltag capabilities
      responses:
        "200": { description: Profile payload }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /handovers/{id}/ack/:
    post:
      summary: Acknowledge the explicitly read revision; idempotent per user and revision
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Existing acknowledgement returned unchanged
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ok: { type: boolean }
                  - { $ref: "#/components/schemas/HandoverAck" }
        "201":
          description: Acknowledgement created
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ok: { type: boolean }
                  - { $ref: "#/components/schemas/HandoverAck" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /defects/:
    get:
      summary: List station defects
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: priority, schema: { type: string } }
        - { in: query, name: overdue, schema: { type: boolean } }
      responses:
        "200":
          description: Defect list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/Defect" }
    post:
      summary: Create a defect
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DefectWrite" }
      responses:
        "201": { description: Defect created }
        "422": { $ref: "#/components/responses/Error" }
  /defects/{id}/:
    get:
      summary: Defect detail including events and attachment metadata
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200":
          description: Defect detail including events and attachment metadata
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ok: { type: boolean }
                  - { $ref: "#/components/schemas/DefectDetail" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update owner, due date, priority or descriptive fields
      description: >-
        Only the fields of DefectPatch are applied. title and category cannot be
        changed after creation: they are ignored when sent. A body without at
        least one changeable field is answered with 422, and so is a body that
        contains nothing but ignored fields.
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DefectPatch" }
      responses:
        "200": { description: Defect updated }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /defects/{id}/status/:
    post:
      summary: Change defect status; closing requires shift lead/admin
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, in_progress, waiting, done] }
      responses:
        "200": { description: Defect updated }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /defects/{id}/attachments/:
    get:
      summary: List image attachments for a defect
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Attachment metadata list }
    post:
      summary: Upload JPEG, PNG or WebP up to 2 MiB each; max 8 images and 12 MiB total per defect
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, content_type, data_base64]
              properties:
                filename: { type: string, maxLength: 180 }
                content_type: { type: string, enum: [image/jpeg, image/png, image/webp] }
                data_base64: { type: string, format: byte }
      responses:
        "201": { description: Attachment created }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /attachments/{id}/:
    get:
      summary: Download an authenticated defect image
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Image bytes }
        "404": { $ref: "#/components/responses/Error" }
  /assets/:
    get:
      summary: Vehicle/device status board
      responses:
        "200": { description: Asset list }
    post:
      summary: Create an asset (shift lead/admin)
      responses:
        "201": { description: Asset created }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /inventory/:
    get:
      summary: List keys and pooled devices
      responses:
        "200": { description: Inventory list }
    post:
      summary: Create inventory item (shift lead/admin)
      responses:
        "201": { description: Inventory item created }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /checklisten/:
    get:
      summary: Active checklists including recurring schedule state
      responses:
        "200": { description: Checklist list }
        "404": { $ref: "#/components/responses/Error" }
...
  /checklisten/{id}/abschluss/:
    post:
      summary: Alias for /erledigt/
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "201": { description: Completion recorded }
  /checklisten/{id}/schedule/:
    get:
      summary: Read recurring checklist schedule
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Schedule }
    put:
      summary: Set recurring checklist schedule (shift lead/admin)
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Schedule updated }
    delete:
      summary: Remove recurring schedule (shift lead/admin)
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Schedule removed }
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: Token <secret> or Bearer <secret>"
  parameters:
    IntegerId:
      in: path
      name: id
      required: true
      schema: { type: integer }
  responses:
    Error:
      description: Canonical API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
              enum: [validation_error, auth_required, forbidden, mfa_required, mfa_setup_required, not_found, rate_limit, conflict, server_error]
            message: { type: string }
            correlation_id: { type: string }
    Defect:
      allOf:
        - { $ref: "#/components/schemas/DefectWrite" }
        - type: object
          required: [id, status, created_at, updated_at]
          properties:
            id: { type: integer }
            status: { type: string, enum: [open, in_progress, waiting, done] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            attachment_count: { type: integer }
    DefectWrite:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 160 }
        description: { type: string, maxLength: 3000 }
        asset_ref: { type: string, maxLength: 160 }
        priority: { type: string, enum: [normal, important, urgent] }
        owner: { type: string, description: Username of an active station member }
        due_at: { type: string, format: date-time, nullable: true }
        category: { type: string, enum: [vehicle, material, safety, facility, key, device, task] }
    DefectPatch:
      type: object
      minProperties: 1
      description: >-
        Only these fields can be changed after creation. title and category are
        fixed at creation time; sending them is not an error, but they are
        ignored and do not count towards the required minimum of one field.
        At least one of the listed fields must be present, otherwise the server
        answers 422.
      properties:
        description: { type: string, maxLength: 3000 }
        asset_ref: { type: string, maxLength: 160 }
        priority: { type: string, enum: [normal, important, urgent] }
        owner: { type: string, description: Username of an active station member }
        due_at: { type: string, format: date-time, nullable: true }
    DefectEvent:
      type: object
      properties:
        kind: { type: string }
        from_status: { type: string, nullable: true }
        to_status: { type: string, nullable: true }
        by: { type: string }
        at: { type: string, format: date-time }
    Attachment:
      type: object
      properties:
        id: { type: integer }
        defect_id: { type: integer }
        filename: { type: string }
        content_type: { type: string }
        size: { type: integer }
        created_at: { type: string, format: date-time }
        uploaded_by: { type: string }
        download_url: { type: string }
    DefectDetail:
      allOf:
        - { $ref: "#/components/schemas/Defect" }
        - type: object
          properties:
            events:
              type: array
              description: At most the 100 most recent events.
              items: { $ref: "#/components/schemas/DefectEvent" }
            attachments:
              type: array
              items: { $ref: "#/components/schemas/Attachment" }
...
    HandoverAck:
      type: object
      required: [handover_id, by, at]
      properties:
        handover_id: { type: integer }
        by: { type: string }
        at: { type: string, format: date-time }
        version: { type: integer, nullable: true }
...